parallel search for items and tuning. fix elasticsearch connection settings
